Message insert to table feature
API now can get sender key, reciever key, and a message as parameters and insert them into a table.
If this table doesn't exist, it creates one and inserts the message.
even if sender and reciever positions change and sender becomes reciever, app doesn't create a duplicate of that table and inserts the values into the existing one